Move enums to Python script to reduce library size.
