+++ cython/python_solvespace/slvs.pyi
@@ -1,8 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from typing import Tuple, List, Sequence, Counter, ClassVar
-from enum import IntEnum, auto
-
 
 def quaternion_u(
     qw: float,
@@ -37,52 +35,6 @@ def make_quaternion(
     vz: float
 ) -> Tuple[float, float, float, float]:
     ...
-
-
-class Constraint(IntEnum):
-    # Expose macro of constraint types
-    POINTS_COINCIDENT = 100000
-    PT_PT_DISTANCE = auto()
-    PT_PLANE_DISTANCE = auto()
-    PT_LINE_DISTANCE = auto()
-    PT_FACE_DISTANCE = auto()
-    PT_IN_PLANE = auto()
-    PT_ON_LINE = auto()
-    PT_ON_FACE = auto()
-    EQUAL_LENGTH_LINES = auto()
-    LENGTH_RATIO = auto()
-    EQ_LEN_PT_LINE_D = auto()
-    EQ_PT_LN_DISTANCES = auto()
-    EQUAL_ANGLE = auto()
-    EQUAL_LINE_ARC_LEN = auto()
-    SYMMETRIC = auto()
-    SYMMETRIC_HORIZ = auto()
-    SYMMETRIC_VERT = auto()
-    SYMMETRIC_LINE = auto()
-    AT_MIDPOINT = auto()
-    HORIZONTAL = auto()
-    VERTICAL = auto()
-    DIAMETER = auto()
-    PT_ON_CIRCLE = auto()
-    SAME_ORIENTATION = auto()
-    ANGLE = auto()
-    PARALLEL = auto()
-    PERPENDICULAR = auto()
-    ARC_LINE_TANGENT = auto()
-    CUBIC_LINE_TANGENT = auto()
-    EQUAL_RADIUS = auto()
-    PROJ_PT_DISTANCE = auto()
-    WHERE_DRAGGED = auto()
-    CURVE_CURVE_TANGENT = auto()
-    LENGTH_DIFFERENCE = auto()
-
-
-class ResultFlag(IntEnum):
-    # Expose macro of result flags
-    OKAY = 0
-    INCONSISTENT = auto()
-    DIDNT_CONVERGE = auto()
-    TOO_MANY_UNKNOWNS = auto()
 
 
 class Params:
@@ -178,7 +130,7 @@ class SolverSystem:
     def faileds(self) -> List[int]:
         ...
 
-    def solve(self) -> ResultFlag:
+    def solve(self) -> int:
         ...
 
     def create_2d_base(self) -> Entity:
@@ -219,7 +171,7 @@ class SolverSystem:
 
     def add_constraint(
         self,
-        c_type: Constraint,
+        c_type: int,
         wp: Entity,
         v: float,
         p1: Entity,
